Fix not being able to change update interval
When setting the webcam update interval, the value would be a string.  Using isNumber would evaluate to false since it checks if value is the "number" type, but it has a type of string.  This was causing it to never update the webcamUpdateInterval.

I propose explicitly converting it to an integer, this avoids type coercion later on as well.