Added nextTick=>return to registerCategory function
If a plugin attempts to register a new menu category using 'registerCategory', and then immediately attempts to register routes for child components of the new menu category, the menu does not update on screen. By adding a nextTick=>return to the 'registerCategory' function, it ensures the DOM/Component is updated before the child components are added, and the menu is updated on screen as expected.